updates based on pull request feedback
