--- src/opera/test/data/test_disp_input/compressed_slc_t087_185678_iw2_20180101_20180210.h5
@@ -1,1 +1,0 @@
-non-empty file
