--- src/opera/test/data/test_disp_input/t087_185678_iw2_amp_dispersion.tif
@@ -1,1 +1,0 @@
-non-empty file
