+++ src/opera/test/pge/disp_s1/test_disp_s1_pge.py
@@ -19,6 +19,7 @@ from pkg_resources import resource_filename
 
 from opera.pge import RunConfig
 from opera.pge.disp_s1.disp_s1_pge import DispS1Executor
+from opera.util.input_validation import validate_disp_inputs
 from opera.util import PgeLogger
 
 
@@ -57,6 +58,14 @@ class DispS1PgeTestCase(unittest.TestCase):
 
         self.input_file = tempfile.NamedTemporaryFile(
             dir=input_dir, prefix="test_input_", suffix=".tiff"
+        )
+
+        os.system(
+            f"echo \"non-empty file\" > {join(input_dir, 'compressed_slc_t087_185678_iw2_20180101_20180210.h5')}"
+        )
+
+        os.system(
+            f"echo \"non-empty file\" > {join(input_dir, 't087_185678_iw2_amp_dispersion.tif')}"
         )
 
         os.chdir(self.working_dir.name)
@@ -248,6 +257,165 @@ class DispS1PgeTestCase(unittest.TestCase):
         finally:
             if exists(test_runconfig_path):
                 os.unlink(test_runconfig_path)
+
+
+    def test_validate_disp_inputs(self):
+        """
+        Test that the validate_disp_inputs function is able to detect non-existent files,
+        zero-size files, and invalid extensions in filenames. Also check that
+        valid files pass validation.
+        """
+
+        # Test non-existent file detection
+        test_filename = 'non_existent_input_file'
+        sas_config = {
+            'runconfig' : {
+                'groups' : {
+                    'input_file_group' : {
+                        'cslc_file_list' : [
+                            test_filename
+                        ]
+                    },
+                    'dynamic_ancillary_file_group' : {
+                    }
+                }
+            }
+        }
+        runconfig = MockRunConfig(sas_config)
+        logger = PgeLogger()
+        with self.assertRaises(RuntimeError):
+            validate_disp_inputs(runconfig, logger, "DISP-S1")
+
+        # Check to see that the RuntimeError is as expected
+        logger.close_log_stream()
+        log_file = logger.get_file_name()
+        self.assertTrue(exists(log_file))
+        with open(log_file, 'r', encoding='utf-8') as lfile:
+            log = lfile.read()
+        self.assertIn(f'Could not locate specified input {test_filename}.', log)
+
+        # Test invalid file extension
+        logger = PgeLogger()
+        test_filename = "test invalid extension.inv"
+        with open(test_filename, 'w') as ief:
+            ief.write("\n")
+        sas_config['runconfig']['groups']['input_file_group']['cslc_file_list'] = [test_filename]
+        runconfig = MockRunConfig(sas_config)
+        with self.assertRaises(RuntimeError):
+            validate_disp_inputs(runconfig, logger, "DISP-S1")
+
+        # Check to see that the RuntimeError is as expected
+        logger.close_log_stream()
+        log_file = logger.get_file_name()
+        self.assertTrue(exists(log_file))
+        with open(log_file, 'r', encoding='utf-8') as lfile:
+            log = lfile.read()
+        self.assertIn(f'Input file {test_filename} does not have an expected file extension', log)
+        os.remove(test_filename)
+
+        # Test for non-zero file size
+        logger = PgeLogger()
+        test_filename = "zero size file.h5"
+        open(test_filename, 'w').close()
+        self.assertTrue(exists(test_filename))
+        sas_config['runconfig']['groups']['input_file_group']['cslc_file_list'] = [test_filename]
+        runconfig = MockRunConfig(sas_config)
+        with self.assertRaises(RuntimeError):
+            validate_disp_inputs(runconfig, logger, "DISP-S1")
+
+        # Check to see that the RuntimeError is as expected
+        logger.close_log_stream()
+        log_file = logger.get_file_name()
+        self.assertTrue(exists(log_file))
+        with open(log_file, 'r', encoding='utf-8') as lfile:
+            log = lfile.read()
+        self.assertIn(f'Input file {test_filename} size is 0. Size must be greater than 0.', log)
+        os.remove(test_filename)
+
+        # Test all input files with valid files
+        cslc_file_list = ['t087_185678_iw2_20180222_VV.h5', 't087_185687_iw1_20180716_VV.h5']
+        for f in cslc_file_list:
+            with open(f, 'w') as wf:
+                wf.write('\n')
+        sas_config['runconfig']['groups']['input_file_group']['cslc_file_list'] = cslc_file_list
+
+        sas_config['runconfig']['groups']['dynamic_ancillary_file_group'] = {}
+        amplitude_dispersion_files = ['t087_185678_iw2_amp_dispersion.tif', 't087_185687_iw1_amp_dispersion.tif']
+        for f in amplitude_dispersion_files:
+            with open(f, 'w') as wf:
+                wf.write('\n')
+        sas_config['runconfig']['groups']['dynamic_ancillary_file_group']['amplitude_dispersion_files'] = amplitude_dispersion_files
+
+        amplitude_mean_files = ['t087_185678_iw2_amp_mean.tif', 't087_185687_iw1_amp_mean.tif']
+        for f in amplitude_mean_files:
+            with open(f, 'w') as wf:
+                wf.write('\n')
+        sas_config['runconfig']['groups']['dynamic_ancillary_file_group']['amplitude_mean_files'] = amplitude_mean_files
+
+        geometry_files = ['t087_185678_iw2_topo.h5', 't087_185687_iw1_topo.h5']
+        for f in geometry_files:
+            with open(f, 'w') as wf:
+                wf.write('\n')
+        sas_config['runconfig']['groups']['dynamic_ancillary_file_group']['geometry_files'] = geometry_files
+
+        mask_files = ['water_mask.tif']
+        for f in mask_files:
+            with open(f, 'w') as wf:
+                wf.write('\n')
+        sas_config['runconfig']['groups']['dynamic_ancillary_file_group']['mask_file'] = mask_files
+
+        dem_file = 'dem.tif'
+        with open(dem_file, 'w') as wf:
+            wf.write('\n')
+        sas_config['runconfig']['groups']['dynamic_ancillary_file_group']['dem_file'] = dem_file
+
+        tec_files = ['jplg0410.18i.Z', 'jplg1970.18i.Z']
+        for f in tec_files:
+            with open(f, 'w') as wf:
+                wf.write('\n')
+        sas_config['runconfig']['groups']['dynamic_ancillary_file_group']['tec_files'] = tec_files
+
+        weather_model_files = ['GMAO_tropo_20180210T000000_ztd.nc', 'GMAO_tropo_20180716T000000_ztd.nc']
+        for f in weather_model_files:
+            with open(f, 'w') as wf:
+                wf.write('\n')
+        sas_config['runconfig']['groups']['dynamic_ancillary_file_group']['weather_model_files'] = weather_model_files
+
+        logger = PgeLogger()
+        runconfig = MockRunConfig(sas_config)
+        validate_disp_inputs(runconfig, logger, "DISP-S1")
+
+        logger.close_log_stream()
+        log_file = logger.get_file_name()
+        self.assertTrue(exists(log_file))
+        with open(log_file, 'r', encoding='utf-8') as lfile:
+            log = lfile.read()
+        print("begin log contents")
+        print(log)
+        print("end log contents")
+        self.assertIn('overall.log_messages.debug: 0', log)
+        self.assertIn('overall.log_messages.info: 0', log)
+        self.assertIn('overall.log_messages.warning: 0', log)
+        self.assertIn('overall.log_messages.critical: 0', log)
+
+        files_to_remove = (cslc_file_list + amplitude_dispersion_files +
+                          amplitude_mean_files + geometry_files + mask_files +
+                          tec_files + weather_model_files)
+        files_to_remove.append(dem_file)
+        for f in files_to_remove:
+            os.remove(f)
+
+
+class MockRunConfig:
+    """Mock runconfig for testing"""
+
+    def __init__(self, sas_config):
+        self._sas_config_dict = sas_config
+
+    @property
+    def sas_config(self):
+        return self._sas_config_dict
+
 
 if __name__ == "__main__":
     unittest.main()
